tp/requirements.txt
bandit==1.8.3
fastapi==0.115.12
filelock==3.18.0
pip_audit==2.9.0
python-dotenv==1.1.0
uvicorn==0.34.3
